requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.0.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
phonenumbers==8.13.42
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
razorpay==1.4.2
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
